requirements.txt
Flask
jinja2
psycopg2-binary
sqlalchemy
gunicorn